# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "UIP Smart City API",
    "description": "REST API for UIP Smart City Platform — Sprint 1-3",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/forecast/energy": {
      "get": {
        "tags": [
          "forecast"
        ],
        "summary": "Get energy forecast for a building",
        "description": "Returns ARIMA energy forecast with 95% confidence intervals. Tenant ID extracted from JWT.",
        "operationId": "forecastEnergy",
        "parameters": [
          {
            "name": "buildingId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Building ID"
          },
          {
            "name": "horizonDays",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 90,
              "default": 30
            },
            "description": "Forecast horizon in days (1-90)"
          }
        ],
        "responses": {
          "200": {
            "description": "Forecast result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter or invalid tenant"
          },
          "403": {
            "description": "Missing tenant context"
          },
          "503": {
            "description": "Forecast service unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ForecastResponse": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string"
          },
          "buildingId": {
            "type": "string"
          },
          "model": {
            "type": "string",
            "enum": [
              "ARIMA",
              "NAIVE",
              "NONE"
            ]
          },
          "isFallback": {
            "type": "boolean"
          },
          "mape": {
            "type": "number",
            "nullable": true,
            "description": "Mean Absolute Percentage Error"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ForecastPoint"
            }
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ForecastPoint": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "actualValue": {
            "type": "number",
            "nullable": true
          },
          "predictedValue": {
            "type": "number"
          },
          "confidenceUpper": {
            "type": "number",
            "description": "95% CI upper bound"
          },
          "confidenceLower": {
            "type": "number",
            "description": "95% CI lower bound"
          },
          "isAnomaly": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}